requirements.txt
langchain
langchain-chroma
langchain-huggingface
langchain-groq
langchain-text-splitters
pymupdf
pymupdf4llm
python-dotenv
chromadb
sentence-transformers